+++ packages/form-render/src/widgets/antd/upload.js
@@ -1,6 +1,6 @@
 import { UploadOutlined } from '@ant-design/icons';
 import { Button, message, Upload } from 'antd';
-import { get } from 'lodash-es'
+import { get } from 'lodash-es';
 import React from 'react';
 
 export default function FrUpload({
@@ -9,7 +9,7 @@ export default function FrUpload({
   onChange,
   uploadProps,
   buttonProps,
-  schema
+  schema,
 }) {
   const props = {
     name: 'file',
@@ -18,8 +18,10 @@ export default function FrUpload({
     onChange(info) {
       if (info.file.status === 'done') {
         message.success(`${info.file.name} 上传成功`);
-        const path = get(schema, 'props.path', '')
-        const url = path ? get(info.file.response, path) : info.file.response.url;
+        const path = get(schema, 'props.path', '');
+        const url = path
+          ? get(info.file.response, path)
+          : info.file.response.url;
         onChange(url);
       } else if (info.file.status === 'error') {
         message.error(`${info.file.name} 上传失败`);